laptop_rental/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
black==24.4.2
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contexttimer==0.3.3
cryptography==42.0.8
DateTime==5.5
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.1.0
django-paypal==2.1
django-redis==5.4.0
djangorestframework==3.15.1
executing==2.0.1
Faker==25.3.0
filelock==3.13.1
flake8==7.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==22.0.0
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mock==5.1.0
myapp==0.1.dev0
mypy-extensions==1.0.0
numpy==1.26.4
nvm==1.0.21
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
postgres==4.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pypng==0.20220715.0
pyproject-api==1.6.1
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.6
render==1.0.0
requests==2.32.3
review==1510
singleton-decorator==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
testcase==0.1.0
tox==4.13.0
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.2
whitenoise==6.7.0
zope.interface==6.4.post2
